AM modulation with noise signal
Simulate an amplitude modulation (AM) system with all input, carrier and output signals. Say the input signal is a cosine wave with amplitude as 2V and frequency as 1000Hz. The carrier signal is also a cosine wave with amplitude as 5V and frequency as 10KHz. The modulation degree is 0.5, and the initial phases of all cosine wave are 0. consider an actual vivid simulation mode, and add random noise in output signal. In this simulation, we divide time domain into several duration, and call each duration as frame. The scanning cycle of an oscilloscope is equal to frame period, that means each time we simulate a frame of signal, and the display will be refreshed once. Therefore, we can get a constantly sliding input signal, a carrier signal with phase jitter, and output signal with noise.