sample mean comparison with whole mean
Write a script that calculates the mean of five samples of data from a vector of random data. Calculate the overall mean. Use a for loop to perform the calculations. For each iteration of the loop print out the intermediate results. Use an if..else control block to display the results depending on whether the mean of the samples is less than, greater than or equal to the overall mean.